Move definitions out of tlschema.
