+++ packages/utils/src/index.ts
@@ -76,6 +76,7 @@ export {
 export { fpsThrottle, throttleToNextFrame } from './lib/throttle'
 export { Timers } from './lib/timers'
 export type { Expand, RecursivePartial, Required } from './lib/types'
+export { safeParseUrl } from './lib/url'
 export {
 	STRUCTURED_CLONE_OBJECT_PROTOTYPE,
 	isDefined,
