+++ packages/utils/src/lib/url.ts
@@ -1,0 +1,8 @@
+/** @public */
+export const safeParseUrl = (url: string) => {
+	try {
+		return new URL(url)
+	} catch (err) {
+		return
+	}
+}
